+++ src/ray/common/task/task_common.h
@@ -1,7 +1,7 @@
 #ifndef RAY_COMMON_TASK_COMMON_H
 #define RAY_COMMON_TASK_COMMON_H
 
-#include "ray/protobuf/common.pb.h"
+#include "src/ray/protobuf/common.pb.h"
 
 namespace ray {
