+++ src/ray/common/task/task_spec.h
@@ -6,13 +6,13 @@
 #include <unordered_map>
 #include <vector>
 
-#include "ray/common/id.h"
-#include "ray/common/scheduling_resources.h"
-#include "ray/common/task_common.h"
-#include "ray/common/grpc_util.h"
+#include "src/ray/common/id.h"
+#include "src/ray/common/task/scheduling_resources.h"
+#include "src/ray/common/task/task_common.h"
+#include "src/ray/common/grpc_util.h"
 
 extern "C" {
-#include "ray/thirdparty/sha256.h"
+#include "src/ray/thirdparty/sha256.h"
 }
 
 namespace ray {
