+++ src/ray/raylet/node_manager.h
@@ -7,10 +7,10 @@
 #include "ray/rpc/client_call.h"
 #include "ray/rpc/node_manager/node_manager_server.h"
 #include "ray/rpc/node_manager/node_manager_client.h"
-#include "ray/common/task.h"
+#include "ray/common/task/task.h"
 #include "ray/common/client_connection.h"
-#include "ray/common/task_common.h"
-#include "ray/common/scheduling_resources.h"
+#include "ray/common/task/task_common.h"
+#include "ray/common/task/scheduling_resources.h"
 #include "ray/object_manager/object_manager.h"
 #include "ray/raylet/actor_registration.h"
 #include "ray/raylet/lineage_cache.h"
