+++ src/ray/raylet/raylet_client.cc
@@ -14,7 +14,7 @@
 
 #include "ray/common/common_protocol.h"
 #include "ray/common/ray_config.h"
-#include "ray/common/task_spec.h"
+#include "ray/common/task/task_spec.h"
 #include "ray/raylet/format/node_manager_generated.h"
 #include "ray/util/logging.h"
