+++ src/ray/raylet/worker_pool.h
@@ -8,8 +8,8 @@
 
 #include "ray/common/client_connection.h"
 #include "ray/gcs/client.h"
-#include "ray/common/task_common.h"
-#include "ray/common/task.h"
+#include "ray/common/task/task_common.h"
+#include "ray/common/task/task.h"
 #include "ray/raylet/worker.h"
 
 namespace ray {
